speculative_decoding/requirements.txt
black==24.10.0
huggingface-hub==0.26.2
llama_cpp_python==0.3.1
